refactor - renamed shared TextField component file
