+++ src/pages/auth/login.tsx
@@ -6,7 +6,7 @@ import Router from "next/router";
 import { useEffect } from "react";
 import { LOGIN_MUTATION } from "../../client/auth/mutations";
 import { isLoggedInVar, isNavDrawerOpenVar } from "../../client/cache";
-import { TextField } from "../../components/Shared/Field";
+import { TextField } from "../../components/Shared/TextField";
 import Flex from "../../components/Shared/Flex";
 import LevelOneHeading from "../../components/Shared/LevelOneHeading";
 import ProgressBar from "../../components/Shared/ProgressBar";
